requirements.txt
streamlit
pandas
openai
python-dotenv
plotly
sqlalchemy
tabulate